services/gpu-demo/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.9.2

# ML and GPU
torch==2.5.1
torchvision==0.20.1
ultralytics==8.3.38
numpy==1.26.4
pillow==11.0.0
opencv-python-headless==4.10.0.84

# GPU monitoring (optional)
nvidia-ml-py==12.560.30

# Observability
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-exporter-otlp==1.28.2
prometheus-client==0.21.0
structlog==24.4.0

# Utils
python-multipart==0.0.18
httpx==0.28.0
python-json-logger==2.0.7
